Move SEO section to components folder
